+++ CHANGELOG.md
@@ -18,3 +18,11 @@
 - `Agent Julia` model can now loop simulations in the same manner as the
   Gillespie SSA model. Because of the complexity of the model, it does take much
   longer to complete the simulation at `n = 1000`.
+- `Agent Julia` simulation files now dump their output within an organised
+  directory structure.
+- Updated `integrated_gpu_support.sh` to take the name of a simulation file as
+  an argument.
+- All model files now show an error informing the user that they are meant to be
+  ran using a simulation file and not on their own.
+- Examples of both models are composites from 1000 loops each. Note this takes
+  significantly longer for the `Agent Julia` model.